Modify how legacy page works
