+++ frontend/src/pages/Main/Home/Checklist/constants.js
@@ -1,0 +1,58 @@
+export const CHECKLIST_STORAGE_KEY = "anythingllm_checklist_completed";
+export const CHECKLIST_HIDDEN = "anythingllm_checklist_dismissed";
+export const CHECKLIST_ITEMS = [
+  {
+    id: "create_workspace",
+    title: "Create a workspace",
+    description: "Create your first workspace to get started",
+    action: "Create",
+    handler: "createWorkspace",
+    completed: false,
+    roles: ["admin", "manager"],
+  },
+  {
+    id: "send_chat",
+    title: "Send a chat",
+    description: "Start a conversation with your AI assistant",
+    action: "Chat",
+    handler: "sendChat",
+    completed: false,
+    roles: ["admin", "manager", "default"],
+  },
+  {
+    id: "embed_document",
+    title: "Embed a document",
+    description: "Add your first document to your workspace",
+    action: "Embed",
+    handler: "embedDocument",
+    completed: false,
+    roles: ["admin", "manager"],
+  },
+  {
+    id: "setup_system_prompt",
+    title: "Set up a system prompt",
+    description: "Configure your AI assistant's behavior",
+    action: "Set Up",
+    handler: "setSystemPrompt",
+    completed: false,
+    roles: ["admin", "manager"],
+  },
+  {
+    id: "define_slash_command",
+    title: "Define a slash command",
+    description: "Create custom commands for your assistant",
+    action: "Define",
+    handler: "setSlashCommand",
+    completed: false,
+    roles: ["admin", "manager", "default"],
+  },
+  {
+    id: "visit_community",
+    title: "Visit Community Hub",
+    description: "Explore community resources and templates",
+    action: "Browse",
+    handler: "visitCommunityHub",
+    completed: false,
+    roles: ["admin", "manager", "default"],
+  },
+];
